requirements.txt
boto==2.49.0
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
stripe==2.10.1
urllib3==1.23
